requirements.txt
Flask>=2.0
requests>=2.28
youtube-transcript-api>=0.5.0
python-dotenv>=1.0.0
reportlab>=4.0
flask-cors>=4.0
google-generativeai>=0.4.0
